requirements.txt
Flask==3.0.0
pdf2image==1.16.3
Pillow==10.1.0
pytesseract==0.3.10